% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaggeredSynthDiD.R
\name{StaggeredSynthDiD}
\alias{StaggeredSynthDiD}
\title{Synthetic DiD estimation for staggered adoption settings}
\usage{
StaggeredSynthDiD(
  data,
  unit = "Unit",
  time = "Time",
  outcome = "Y",
  treatment = "W",
  covariates = NULL,
  vcov = "jack",
  iterations = 100,
  ci_cov = 0.95
)
}
\arguments{
\item{data}{The data set. Must be in long format, i.e., one row per unit time combination.}

\item{unit}{The name of the column containing the unit id.}

\item{time}{The name of the column containing the time period.}

\item{outcome}{The name of the column containing the outcome.}

\item{treatment}{The name of the column containing the treatment indicator.}

\item{covariates}{A character vector of covariate column names. NULL if no covariates should be included.}

\item{vcov}{The variance estimator to be used. Should be one of "boot", "jack" or "placebo".}

\item{iterations}{Number of resampling iterations. Only relevant for the bootstrap and placebo variance estimators.}

\item{ci_cov}{Desired nominal coverage probability of the confidence interval.}
}
\value{
A list containing the point estimate, the standard error, and a confidence interval with user-specified nominal coverage.
}
\description{
Estimates the Average Treatment Effect (ATE) in staggered adoption settings. Standard errors and a confidence interval with user-specified desired coverage is provided.
}
\references{
Arkhangelsky, Dmitry, Susan Athey, David A. Hirshberg, Guido W. Imbens, and Stefan Wager. 2021. "Synthetic Difference-in-Differences." American Economic Review, 111 (12): 4088-4118.
}
